zepimg_111/requirements.txt
fastapi
uvicorn
python-dotenv
Pillow
google-generativeai
supabase
cryptography
pyasn1
pyasn1-modules
